+++ ml_tools/_core/__init__.py
@@ -1,7 +1,9 @@
 from ._script_info import _imprimir_disponibles
 from ._logger import get_logger
+from ._helpers import wrap_text
 
 __all__ = [
     "_imprimir_disponibles",
-    "get_logger"
+    "get_logger",
+    "wrap_text"
 ]